DOCKER-65 SIGTERM is not propogated to the DXP process so Graceful shutdown is broken
Prevent the entrypoint from forking on bash SHEBANG and breaking the signal handling meant to be on PID 1
With suggestions from @thecosmicfrog

+++ templates/base/scripts/liferay_entrypoint.sh
@@ -3,7 +3,11 @@
 source /usr/local/bin/_liferay_common.sh
 
 function handle_kill_TERM {
-	kill -TERM ${START_LIFERAY_PID}
+	exec 5<>/dev/tcp/localhost/8005 && echo "SHUTDOWN" >&5
+
+	if [ $? -gt 0 ]; then
+		kill -TERM ${START_LIFERAY_PID}
+	fi
 }
 
 function main {
